+++ src/tx-data/declare-candidacy.js
@@ -14,6 +14,7 @@ class TxDataDeclareCandidacy {
             {
                 name: 'publicKey',
                 allowZero: true,
+                length: 32,
                 default: Buffer.from([]),
             },
             {