+++ src/tx-data/delegate.js
@@ -8,6 +8,7 @@ class TxDataDelegate {
             {
                 name: 'publicKey',
                 allowZero: true,
+                length: 32,
                 default: Buffer.from([]),
             },
             {