+++ src/tx-data/edit-candidate.js
@@ -8,6 +8,13 @@ class TxDataDeclareCandidacy {
             {
                 name: 'publicKey',
                 allowZero: true,
+                length: 32,
+                default: Buffer.from([]),
+            },
+            {
+                name: 'newPublicKey',
+                allowZero: true,
+                length: 32,
                 default: Buffer.from([]),
             },
             {