+++ src/tx-data/set-candidate-on.js
@@ -8,6 +8,7 @@ class TxDataSetCandidateOn {
             {
                 name: 'publicKey',
                 allowZero: true,
+                length: 32,
                 default: Buffer.from([]),
             }];